requirements.txt
# VENOM Λ-Genesis Python Dependencies

# Core dependencies
# python>=3.8

# Web framework (REST API)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# gRPC
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1

# Numerical computing
numpy==1.26.2

# Vector database (optional - for RAG)
faiss-cpu==1.7.4  # Uncomment if using FAISS

# Data processing
pandas==2.1.3

# Logging and utilities
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.11.0
pylint==3.0.3
mypy==1.7.1

# Optional: ML/AI libraries
# tensorflow-lite==2.14.0  # If running TFLite in Python
# torch==2.1.1  # If using PyTorch
# transformers==4.35.2  # For LLM integration

# Optional: Knowledge processing
# sentence-transformers==2.2.2  # For embeddings
# chromadb==0.4.18  # Alternative vector DB

# Optional: Data validation
# marshmallow==3.20.1

# Monitoring and profiling
# psutil==5.9.6  # System monitoring
# py-spy==0.3.14  # Performance profiling
